phase3: extract types, utils, and components from WorkoutDashboard
- Extract WorkoutSession, WorkoutApiSession, Goal, GoalProgress types to src/types/workout.ts
- Extract goal API helpers (fetchGoalsFromAPI, saveGoalToAPI, saveGoalsToStorage, loadGoalsFromStorage) to src/utils/goalsApi.ts
- Extract OCR parsing (ocrNormalize, parseTonalOCR) to src/utils/tonalOCR.ts
- Extract SettingsModal to src/components/SettingsModal.tsx (144 lines)
- Extract DashboardHeader to src/components/DashboardHeader.tsx (209 lines, mobile + desktop)
- Update all 10 consumer files to import types from @/types/workout
- Re-export types from WorkoutDashboard for any remaining compatibility
- WorkoutDashboard reduced from 1,316 to 734 lines
