+++ src/components/GoalTracker.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useEffect } from 'react'
 import { Dumbbell, Activity, CheckCircle, Calendar, BarChart3 } from 'lucide-react'
-import { WorkoutSession } from './WorkoutDashboard'
+import { WorkoutSession } from '@/types/workout'
 import { formatNumber } from '../utils/numberFormat'
 
 interface GoalTrackerProps {
