+++ src/components/WorkoutDashboard.tsx
@@ -8,311 +8,17 @@ import { ClientProgressChart } from './ClientProgressChart'
 import { GoalTracker } from './GoalTracker'
 import { GoalModal } from './GoalModal'
 import { WorkoutSummary } from './WorkoutSummary'
-import { ThemeToggle } from './ThemeToggle'
-import { AuthHeader } from './AuthHeader'
-import { Plus, X, Target, Calendar, Settings, RefreshCw, Upload, Check } from 'lucide-react'
+import { DashboardHeader } from './DashboardHeader'
+import { SettingsModal } from './SettingsModal'
+import { Plus, X, Target, Check } from 'lucide-react'
 import { applyDefaultMileage, applyWorkoutMultipliers } from '../utils/workoutMultipliers'
 import { useSettings } from '../context/SettingsContext'
+import { parseTonalOCR } from '../utils/tonalOCR'
+import { fetchGoalsFromAPI, saveGoalToAPI, saveGoalsToStorage } from '../utils/goalsApi'
+import type { WorkoutSession, WorkoutApiSession, Goal } from '@/types/workout'
 
-export interface WorkoutSession {
-  id: string
-  date: Date
-  source: string
-  activity: string
-  minutes: number
-  miles?: number
-  adjustedMiles?: number // Miles with multiplier applied (for Cannondale)
-  adjustedMinutes?: number // Minutes with multiplier applied (for Cannondale)
-  estimatedMiles?: boolean // True when miles were calculated from default cycling speed
-  weightLifted?: number
-  notes?: string
-}
-
-// Type for API payload returned by /api/workouts
-interface WorkoutApiSession {
-  id: string
-  date: string
-  source: string
-  activity: string
-  minutes: number
-  miles?: number | null
-  weightLifted?: number | null
-  notes?: string | null
-}
-
-export interface Goal {
-  id: string
-  name: string
-  year: number
-  
-  // Annual targets
-  annualWeightTarget: number // Total lbs for the year
-  minutesPerSession: number // Minutes per individual session
-  weeklySessionsTarget: number // Sessions per week
-  
-  // Calculated fields (derived from above)
-  weeklyMinutesTarget: number // minutesPerSession * weeklySessionsTarget
-  annualMinutesTarget: number // weeklyMinutesTarget * 52
-  quarterlyWeightTarget: number // annualWeightTarget / 4
-  quarterlyMinutesTarget: number // annualMinutesTarget / 4
-  quarterlySessionsTarget: number // weeklySessionsTarget * 13
-  
-  createdAt: Date
-  updatedAt: Date
-}
-
-export interface GoalProgress {
-  // Current period progress
-  currentQuarter: number
-  currentYear: number
-  
-  // Actual progress
-  actualWeightLifted: {
-    quarterToDate: number
-    yearToDate: number
-  }
-  actualMinutes: {
-    quarterToDate: number
-    yearToDate: number
-  }
-  actualSessions: {
-    quarterToDate: number
-    yearToDate: number
-  }
-  
-  // Expected progress (linear)
-  expectedWeightLifted: {
-    quarterToDate: number
-    yearToDate: number
-  }
-  expectedMinutes: {
-    quarterToDate: number
-    yearToDate: number
-  }
-  expectedSessions: {
-    quarterToDate: number
-    yearToDate: number
-  }
-  
-  // Sessions needed
-  sessionsNeededForQuarter: number
-  sessionsNeededForYear: number
-  
-  // Time remaining
-  daysRemainingInQuarter: number
-}
-
-// API helpers for goals with localStorage fallback
-const GOALS_STORAGE_KEY = 'fitness-tracker-goals'
-
-const saveGoalsToStorage = (goals: Goal[]) => {
-  try {
-    if (typeof window === 'undefined') return
-    localStorage.setItem(GOALS_STORAGE_KEY, JSON.stringify(goals))
-  } catch (error) {
-    console.error('Error saving goals to localStorage:', error)
-  }
-}
-
-const loadGoalsFromStorage = (): Goal[] => {
-  try {
-    if (typeof window === 'undefined') return []
-    const stored = localStorage.getItem(GOALS_STORAGE_KEY)
-    if (stored) {
-      const parsed = JSON.parse(stored)
-      // Convert date strings back to Date objects
-      return parsed.map((goal: Omit<Goal, 'createdAt' | 'updatedAt'> & { createdAt: string; updatedAt: string }) => ({
-        ...goal,
-        createdAt: new Date(goal.createdAt),
-        updatedAt: new Date(goal.updatedAt)
-      }))
-    }
-  } catch (error) {
-    console.error('Error loading goals from localStorage:', error)
-  }
-  return []
-}
-
-// API functions for cross-device persistence
-const fetchGoalsFromAPI = async (): Promise<Goal[]> => {
-  try {
-    const response = await fetch('/api/goals')
-    if (!response.ok) throw new Error('Failed to fetch goals')
-    const data = await response.json()
-    return data.goals.map((goal: Goal & { createdAt: string; updatedAt: string }) => ({
-      ...goal,
-      createdAt: new Date(goal.createdAt),
-      updatedAt: new Date(goal.updatedAt)
-    }))
-  } catch (error) {
-    console.error('Error fetching goals from API:', error)
-    // Fallback to localStorage
-    return loadGoalsFromStorage()
-  }
-}
-
-const saveGoalToAPI = async (goalData: Omit<Goal, 'id' | 'createdAt' | 'updatedAt'>, existingId?: string): Promise<Goal | null> => {
-  try {
-    const url = '/api/goals'
-    const method = existingId ? 'PUT' : 'POST'
-    const body = existingId ? { ...goalData, id: existingId } : goalData
-    
-    const response = await fetch(url, {
-      method,
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(body)
-    })
-    
-    if (!response.ok) throw new Error('Failed to save goal')
-    const data = await response.json()
-    return {
-      ...data.goal,
-      createdAt: new Date(data.goal.createdAt),
-      updatedAt: new Date(data.goal.updatedAt)
-    }
-  } catch (error) {
-    console.error('Error saving goal to API:', error)
-    return null
-  }
-}
-
-// Normalize common OCR character misreads (letter↔digit swaps)
-function ocrNormalize(text: string): string {
-  return text
-    .replace(/[|]/g, '/')       // pipe → slash (date separators)
-    .replace(/l(?=\d)/g, '1')   // lowercase L before digit → 1
-    .replace(/(?<=\d)l/g, '1')  // lowercase L after digit → 1
-    .replace(/O(?=\d)/g, '0')   // capital O before digit → 0
-    .replace(/(?<=\d)O/g, '0')  // capital O after digit → 0
-    .replace(/(?<=\d)\s*[)\]]/g, '') // stray brackets after digits
-    .replace(/(\d,)\s+(\d)/g, '$1$2') // collapse spaces in comma-formatted numbers (e.g. "12, 379" → "12,379")
-}
-
-function parseTonalOCR(text: string) {
-  // Work with both raw and OCR-normalized text
-  const normalized = ocrNormalize(text)
-  const lines = text.split('\n').map(l => l.trim()).filter(l => l.length > 0)
-
-  // Volume: number followed by "lbs" — OCR may render comma as period, space, or drop it
-  // OCR commonly misreads "lbs" as "1bs", "1s", "Ibs", "ls", etc.
-  const volumeMatch = normalized.match(/(\d{1,3}[,.]\d{3,})\s*(?:lbs|1bs|1s|Ibs|ibs|ls)\b/i)
-    || normalized.match(/(\d{4,})\s*(?:lbs|1bs|1s|Ibs|ibs|ls)\b/i)
-  let weightLifted: number | null = null
-  if (volumeMatch) {
-    weightLifted = parseInt(volumeMatch[1].replace(/[^\d]/g, ''), 10)
-    if (isNaN(weightLifted) || weightLifted === 0) weightLifted = null
-  }
-  // Fallback: look for a large number (4+ digits with comma) on the same line as a MM:SS pattern
-  if (!weightLifted) {
-    const fallbackMatch = normalized.match(/(\d{1,3},\d{3})\s*\S*\s*[-—]?\s*\d{1,3}:\d{2}/)
-    if (fallbackMatch) {
-      weightLifted = parseInt(fallbackMatch[1].replace(/[^\d]/g, ''), 10)
-      if (isNaN(weightLifted) || weightLifted === 0) weightLifted = null
-    }
-  }
-
-  // Duration: MM:SS pattern (strict)
-  const durationMatch = normalized.match(/(\d{1,3}):(\d{2})/)
-  let minutes: number | null = null
-  if (durationMatch) {
-    minutes = parseInt(durationMatch[1], 10) + (parseInt(durationMatch[2], 10) > 0 ? 1 : 0)
-  }
-  // Fallback: minutes-only pattern — digits followed by colon but garbled seconds
-  if (!minutes) {
-    const looseDuration = normalized.match(/(\d{1,3})\s*:\s*(?:\d{0,2}|\D)/)
-    if (looseDuration) {
-      minutes = parseInt(looseDuration[1], 10)
-      if (isNaN(minutes) || minutes === 0) minutes = null
-    }
-  }
-  // Fallback: 4-digit number near VOLUME/DURATION keywords where colon was dropped (e.g. "4205" = 42:05)
-  if (!minutes) {
-    const contextMatch = normalized.match(/\b(\d{3,4})\b(?=\s*(?:VOLUME|DURATION|WEIGHT|MIN))/i)
-      || normalized.match(/(?:VOLUME|DURATION|WEIGHT|lbs)\s+\S*\s*(\d{3,4})\b/i)
-    if (contextMatch) {
-      const raw = contextMatch[1]
-      if (raw.length === 4) {
-        // MMSS: first 2 digits = minutes, last 2 = seconds
-        minutes = parseInt(raw.substring(0, 2), 10) + (parseInt(raw.substring(2), 10) > 0 ? 1 : 0)
-      } else if (raw.length === 3) {
-        // MSS: first digit = minutes, last 2 = seconds
-        minutes = parseInt(raw.substring(0, 1), 10) + (parseInt(raw.substring(1), 10) > 0 ? 1 : 0)
-      }
-      if (minutes !== null && (isNaN(minutes) || minutes === 0 || minutes > 300)) minutes = null
-    }
-  }
-
-  // Date: M/D/YY pattern — try normalized text first (fixes pipe→slash, l→1, O→0)
-  let date: Date | null = null
-  const dateMatch = normalized.match(/(\d{1,2})\/(\d{1,2})\/(\d{2,4})/)
-  if (dateMatch) {
-    const month = parseInt(dateMatch[1], 10)
-    const day = parseInt(dateMatch[2], 10)
-    let year = parseInt(dateMatch[3], 10)
-    if (year < 100) year += 2000
-    if (month >= 1 && month <= 12 && day >= 1 && day <= 31) {
-      date = new Date(year, month - 1, day, 12, 0, 0)
-    }
-  }
-  // Fallback: try matching date-like patterns with OCR artifacts (e.g. "5/4/ 26", "5 /4/26")
-  if (!date) {
-    const looseDateMatch = normalized.match(/(\d{1,2})\s*\/\s*(\d{1,2})\s*\/\s*(\d{2,4})/)
-    if (looseDateMatch) {
-      const month = parseInt(looseDateMatch[1], 10)
-      const day = parseInt(looseDateMatch[2], 10)
-      let year = parseInt(looseDateMatch[3], 10)
-      if (year < 100) year += 2000
-      if (month >= 1 && month <= 12 && day >= 1 && day <= 31) {
-        date = new Date(year, month - 1, day, 12, 0, 0)
-      }
-    }
-  }
-  // Fallback: look for "Weeks" or week-related text with nearby numbers (e.g. "el Weeks to" → date context)
-  // Try dot-separated dates (e.g. "5.4.26")
-  if (!date) {
-    const dotDateMatch = normalized.match(/(\d{1,2})\.(\d{1,2})\.(\d{2,4})/)
-    if (dotDateMatch) {
-      const month = parseInt(dotDateMatch[1], 10)
-      const day = parseInt(dotDateMatch[2], 10)
-      let year = parseInt(dotDateMatch[3], 10)
-      if (year < 100) year += 2000
-      if (month >= 1 && month <= 12 && day >= 1 && day <= 31) {
-        date = new Date(year, month - 1, day, 12, 0, 0)
-      }
-    }
-  }
-  // Fallback: dash-separated dates (e.g. "5-4-26")
-  if (!date) {
-    const dashDateMatch = normalized.match(/(\d{1,2})-(\d{1,2})-(\d{2,4})/)
-    if (dashDateMatch) {
-      const month = parseInt(dashDateMatch[1], 10)
-      const day = parseInt(dashDateMatch[2], 10)
-      let year = parseInt(dashDateMatch[3], 10)
-      if (year < 100) year += 2000
-      if (month >= 1 && month <= 12 && day >= 1 && day <= 31) {
-        date = new Date(year, month - 1, day, 12, 0, 0)
-      }
-    }
-  }
-
-  // Workout name: first meaningful line
-  const workoutName = lines.find(l =>
-    l.length > 3 &&
-    !/^TONAL$/i.test(l) &&
-    !/^VOLUME$/i.test(l) &&
-    !/^DURATION$/i.test(l) &&
-    !/^\d/.test(l) &&
-    !l.includes('|') &&
-    !l.includes('lbs')
-  ) || null
-
-  // Coach/details line with bullet separators
-  const detailsLine = lines.find(l => l.includes('•') || l.includes('·')) || null
-
-  const notes = [workoutName, detailsLine].filter(Boolean).join(' — ')
-
-  return { weightLifted, minutes, date, notes }
-}
+// Re-export types for backward compatibility with existing imports
+export type { WorkoutSession, Goal, GoalProgress } from '@/types/workout'
 
 export function WorkoutDashboard() {
   const { settings, updateSettings } = useSettings()
@@ -797,184 +503,19 @@ export function WorkoutDashboard() {
         onChange={handleTonalImport}
       />
       {/* Header */}
-      <header className="bg-white dark:bg-gray-900 shadow-sm border-b border-gray-200 dark:border-gray-700 sticky top-0 z-50">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          {/* Mobile Layout - Stack vertically */}
-          <div className="sm:hidden py-2">
-            <div className="flex items-center justify-between mb-2">
-              <div className="flex items-center gap-2">
-                {/* eslint-disable-next-line @next/next/no-img-element */}
-                <img
-                  src="/fitness-logo.svg"
-                  alt="Fitness Tracker"
-                  className="w-8 h-8 rounded-lg shadow-sm"
-                />
-                <h1 className="text-base font-bold text-gray-900 dark:text-gray-100">
-                  Fitness Tracker
-                </h1>
-              </div>
-              <div className="flex items-center gap-1.5">
-                <button
-                  onClick={() => handleSync('peloton')}
-                  disabled={syncing !== null}
-                  className="bg-red-600 hover:bg-red-700 disabled:opacity-50 text-white px-2 py-1.5 rounded-lg font-medium transition-colors flex items-center gap-1 shadow-sm text-xs"
-                >
-                  <RefreshCw className={`w-3.5 h-3.5 ${syncing === 'peloton' ? 'animate-spin' : ''}`} />
-                  <span>{syncing === 'peloton' ? 'Syncing…' : 'Peloton'}</span>
-                </button>
-                <button
-                  onClick={() => handleSync('tonal')}
-                  disabled={syncing !== null}
-                  className="bg-gray-900 dark:bg-gray-200 hover:bg-black dark:hover:bg-white disabled:opacity-50 text-white dark:text-gray-900 px-2 py-1.5 rounded-lg font-medium transition-colors flex items-center gap-1 shadow-sm text-xs"
-                >
-                  <RefreshCw className={`w-3.5 h-3.5 ${syncing === 'tonal' ? 'animate-spin' : ''}`} />
-                  <span>{syncing === 'tonal' ? 'Syncing…' : 'Tonal'}</span>
-                </button>
-                {/* OCR screenshot import — fallback when API sync doesn't capture a workout */}
-                <button
-                  onClick={() => tonalFileRef.current?.click()}
-                  disabled={importing}
-                  className="bg-gray-700 dark:bg-gray-500 hover:bg-gray-800 dark:hover:bg-gray-400 disabled:opacity-50 text-white dark:text-gray-900 p-1.5 rounded-lg transition-colors shadow-sm"
-                  title="Import Tonal screenshot (fallback)"
-                >
-                  <Upload className={`w-3.5 h-3.5 ${importing ? 'animate-pulse' : ''}`} />
-                  {importing && importProgress && <span className="text-[10px]">{importProgress}</span>}
-                </button>
-                <button
-                  onClick={() => setShowSettings(true)}
-                  className="p-1.5 rounded-lg bg-gray-100 dark:bg-gray-800 hover:bg-gray-200 dark:hover:bg-gray-700 transition-colors"
-                  title="Settings"
-                >
-                  <Settings className="w-4 h-4 text-gray-500" />
-                </button>
-                <ThemeToggle />
-              </div>
-            </div>
+      <DashboardHeader
+        currentYear={currentYear}
+        setCurrentYear={setCurrentYear}
+        syncing={syncing}
+        importing={importing}
+        importProgress={importProgress}
+        onSync={handleSync}
+        onImportClick={() => tonalFileRef.current?.click()}
+        onSettingsClick={() => setShowSettings(true)}
+        onAddWorkoutClick={() => setShowAddWorkout(true)}
+      />
 
-            <div className="flex items-center justify-between">
-              <div className="flex items-center bg-gray-100 dark:bg-gray-800 rounded-lg p-1">
-                <button
-                  onClick={() => setCurrentYear(2025)}
-                  className={`px-2 py-1 rounded-md text-sm flex items-center gap-1 transition-colors ${
-                    currentYear === 2025
-                    ? 'bg-white dark:bg-gray-700 shadow-sm text-blue-600 dark:text-blue-400'
-                    : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-300'
-                  }`}
-                >
-                  <Calendar className="w-3 h-3" />
-                  <span>2025</span>
-                </button>
-                <button
-                  onClick={() => setCurrentYear(2026)}
-                  className={`px-2 py-1 rounded-md text-sm flex items-center gap-1 transition-colors ${
-                    currentYear === 2026
-                    ? 'bg-white dark:bg-gray-700 shadow-sm text-blue-600 dark:text-blue-400'
-                    : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-300'
-                  }`}
-                >
-                  <Calendar className="w-3 h-3" />
-                  <span>2026</span>
-                </button>
-              </div>
-              <button
-                onClick={() => setShowAddWorkout(true)}
-                className="bg-primary-500 hover:bg-primary-600 text-white px-3 py-1.5 rounded-lg font-medium transition-colors flex items-center gap-1.5 shadow-sm text-sm"
-              >
-                <Plus className="w-4 h-4" />
-                <span>Add</span>
-              </button>
-            </div>
-          </div>
-
-          {/* Desktop Layout - Horizontal */}
-          <div className="hidden sm:flex items-center justify-between h-16 gap-4">
-            <div className="flex items-center gap-3 shrink-0">
-              {/* eslint-disable-next-line @next/next/no-img-element */}
-              <img
-                src="/fitness-logo.svg"
-                alt="Fitness Tracker"
-                className="w-10 h-10 rounded-lg shadow-sm"
-              />
-              <div>
-                <h1 className="text-xl font-bold text-gray-900 dark:text-gray-100 whitespace-nowrap">
-                  Fitness Tracker
-                </h1>
-                <p className="text-xs text-gray-600 dark:text-gray-400 whitespace-nowrap">Track your progress</p>
-              </div>
-            </div>
-            <div className="flex items-center gap-2">
-              <div className="flex items-center bg-gray-100 dark:bg-gray-800 rounded-lg p-1">
-                <button
-                  onClick={() => setCurrentYear(2025)}
-                  className={`px-2 py-1 rounded-md text-sm flex items-center gap-1 transition-colors ${
-                    currentYear === 2025
-                    ? 'bg-white dark:bg-gray-700 shadow-sm text-blue-600 dark:text-blue-400'
-                    : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-300'
-                  }`}
-                >
-                  <Calendar className="w-4 h-4" />
-                  <span>2025</span>
-                </button>
-                <button
-                  onClick={() => setCurrentYear(2026)}
-                  className={`px-2 py-1 rounded-md text-sm flex items-center gap-1 transition-colors ${
-                    currentYear === 2026
-                    ? 'bg-white dark:bg-gray-700 shadow-sm text-blue-600 dark:text-blue-400'
-                    : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-300'
-                  }`}
-                >
-                  <Calendar className="w-4 h-4" />
-                  <span>2026</span>
-                </button>
-              </div>
-              <button
-                onClick={() => handleSync('peloton')}
-                disabled={syncing !== null}
-                className="bg-red-600 hover:bg-red-700 disabled:opacity-50 text-white px-3 py-1.5 rounded-lg font-medium transition-colors flex items-center gap-1.5 shadow-sm text-sm whitespace-nowrap"
-              >
-                <RefreshCw className={`w-4 h-4 ${syncing === 'peloton' ? 'animate-spin' : ''}`} />
-                <span>{syncing === 'peloton' ? 'Syncing…' : 'Sync Peloton'}</span>
-              </button>
-              <button
-                onClick={() => handleSync('tonal')}
-                disabled={syncing !== null}
-                className="bg-gray-900 dark:bg-gray-200 hover:bg-black dark:hover:bg-white disabled:opacity-50 text-white dark:text-gray-900 px-3 py-1.5 rounded-lg font-medium transition-colors flex items-center gap-1.5 shadow-sm text-sm whitespace-nowrap"
-              >
-                <RefreshCw className={`w-4 h-4 ${syncing === 'tonal' ? 'animate-spin' : ''}`} />
-                <span>{syncing === 'tonal' ? 'Syncing…' : 'Sync Tonal'}</span>
-              </button>
-              {/* OCR screenshot import — fallback when API sync doesn't capture a workout */}
-              <button
-                onClick={() => tonalFileRef.current?.click()}
-                disabled={importing}
-                className="bg-gray-700 dark:bg-gray-500 hover:bg-gray-800 dark:hover:bg-gray-400 disabled:opacity-50 text-white dark:text-gray-900 px-3 py-1.5 rounded-lg font-medium transition-colors flex items-center gap-1.5 shadow-sm text-sm whitespace-nowrap"
-                title="Import Tonal screenshot (fallback)"
-              >
-                <Upload className={`w-4 h-4 ${importing ? 'animate-pulse' : ''}`} />
-                <span>{importing ? `Importing ${importProgress || ''}…` : 'Import Tonal'}</span>
-              </button>
-              <button
-                onClick={() => setShowSettings(true)}
-                className="p-2 rounded-lg bg-gray-100 dark:bg-gray-800 hover:bg-gray-200 dark:hover:bg-gray-700 transition-colors"
-                title="Settings"
-              >
-                <Settings className="w-4 h-4 text-gray-500" />
-              </button>
-              <ThemeToggle />
-              <AuthHeader />
-              <button
-                onClick={() => setShowAddWorkout(true)}
-                className="bg-primary-500 hover:bg-primary-600 text-white px-3 py-1.5 rounded-lg font-medium transition-colors flex items-center gap-1.5 shadow-sm text-sm whitespace-nowrap"
-              >
-                <Plus className="w-4 h-4" />
-                <span>Add Workout</span>
-              </button>
-            </div>
-          </div>
-        </div>
-      </header>
-
-      {/* Sync error banner */}
+      {/* Sync success banner */}
       {syncSuccess && (
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 pt-4">
           <div className="bg-green-50 dark:bg-green-900/20 border border-green-200 dark:border-green-800 rounded-lg px-4 py-3 flex items-center justify-between">
@@ -1183,133 +724,10 @@ export function WorkoutDashboard() {
         />
 
         {/* Settings Modal */}
-        {showSettings && (
-          <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center p-4 z-50">
-            <div className="bg-white dark:bg-gray-900 rounded-lg p-6 max-w-md w-full max-h-[90vh] overflow-y-auto">
-              <div className="flex justify-between items-center mb-6">
-                <h2 className="text-xl font-bold text-gray-900 dark:text-white">Settings</h2>
-                <button
-                  onClick={() => setShowSettings(false)}
-                  className="text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200"
-                >
-                  <X className="w-6 h-6" />
-                </button>
-              </div>
-
-              {/* Default Cycling Speed */}
-              <div className="mb-6">
-                <h3 className="text-sm font-semibold text-gray-500 dark:text-gray-400 uppercase tracking-wide mb-3">Default Cycling Speed</h3>
-                <p className="text-sm text-gray-600 dark:text-gray-400 mb-3">
-                  Used to estimate mileage when not recorded by the source.
-                </p>
-                <div className="flex gap-2">
-                  {[17, 18, 19].map((speed) => (
-                    <button
-                      key={speed}
-                      onClick={() => updateSettings({ defaultCyclingSpeed: speed })}
-                      className={`px-3 py-2 rounded-lg text-sm font-semibold transition-colors ${
-                        settings.defaultCyclingSpeed === speed
-                          ? 'bg-indigo-100 dark:bg-indigo-900/40 text-indigo-600 dark:text-indigo-400 ring-2 ring-indigo-500'
-                          : 'bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-400 hover:bg-gray-200 dark:hover:bg-gray-700'
-                      }`}
-                    >
-                      {speed} mph
-                    </button>
-                  ))}
-                </div>
-              </div>
-
-              {/* Outdoor Bonus */}
-              <div className="mb-6">
-                <h3 className="text-sm font-semibold text-gray-500 dark:text-gray-400 uppercase tracking-wide mb-3">Outdoor Bonus</h3>
-                <p className="text-sm text-gray-600 dark:text-gray-400 mb-3">
-                  Cannondale outdoor rides get bonus credit for minutes and miles.
-                </p>
-                <div className="flex gap-2">
-                  {[1, 1.25, 1.5, 1.75, 2].map((m) => (
-                    <button
-                      key={m}
-                      onClick={() => updateSettings({ outdoorMultiplier: m })}
-                      className={`px-3 py-2 rounded-lg text-sm font-semibold transition-colors ${
-                        settings.outdoorMultiplier === m
-                          ? 'bg-indigo-100 dark:bg-indigo-900/40 text-indigo-600 dark:text-indigo-400 ring-2 ring-indigo-500'
-                          : 'bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-400 hover:bg-gray-200 dark:hover:bg-gray-700'
-                      }`}
-                    >
-                      {m}x
-                    </button>
-                  ))}
-                </div>
-              </div>
-
-              {/* Units */}
-              <div className="mb-6">
-                <h3 className="text-sm font-semibold text-gray-500 dark:text-gray-400 uppercase tracking-wide mb-3">Units</h3>
-                <div className="flex gap-2">
-                  <button
-                    onClick={() => updateSettings({ units: 'imperial' })}
-                    className={`flex-1 px-3 py-2 rounded-lg text-sm font-semibold text-center transition-colors ${
-                      settings.units === 'imperial'
-                        ? 'bg-indigo-100 dark:bg-indigo-900/40 text-indigo-600 dark:text-indigo-400 ring-2 ring-indigo-500'
-                        : 'bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-400 hover:bg-gray-200 dark:hover:bg-gray-700'
-                    }`}
-                  >
-                    🇺🇸 Imperial (mi, lbs)
-                  </button>
-                  <button
-                    onClick={() => updateSettings({ units: 'metric' })}
-                    className={`flex-1 px-3 py-2 rounded-lg text-sm font-semibold text-center transition-colors ${
-                      settings.units === 'metric'
-                        ? 'bg-indigo-100 dark:bg-indigo-900/40 text-indigo-600 dark:text-indigo-400 ring-2 ring-indigo-500'
-                        : 'bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-400 hover:bg-gray-200 dark:hover:bg-gray-700'
-                    }`}
-                  >
-                    🌍 Metric (km, kg)
-                  </button>
-                </div>
-              </div>
-
-              {/* Workout Defaults */}
-              <div className="mb-6">
-                <h3 className="text-sm font-semibold text-gray-500 dark:text-gray-400 uppercase tracking-wide mb-3">Default Source</h3>
-                <div className="flex gap-2 flex-wrap">
-                  {['Peloton', 'Tonal', 'Cannondale', 'Other'].map((s) => (
-                    <button
-                      key={s}
-                      onClick={() => updateSettings({ defaultSource: s })}
-                      className={`px-3 py-2 rounded-lg text-sm font-semibold transition-colors ${
-                        settings.defaultSource === s
-                          ? 'bg-indigo-100 dark:bg-indigo-900/40 text-indigo-600 dark:text-indigo-400 ring-2 ring-indigo-500'
-                          : 'bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-400 hover:bg-gray-200 dark:hover:bg-gray-700'
-                      }`}
-                    >
-                      {s}
-                    </button>
-                  ))}
-                </div>
-              </div>
-
-              <div className="mb-2">
-                <h3 className="text-sm font-semibold text-gray-500 dark:text-gray-400 uppercase tracking-wide mb-3">Default Activity</h3>
-                <div className="flex gap-2 flex-wrap">
-                  {['Cycling', 'Weight Lifting', 'Running', 'Walking', 'Yoga', 'Other'].map((a) => (
-                    <button
-                      key={a}
-                      onClick={() => updateSettings({ defaultActivity: a })}
-                      className={`px-3 py-2 rounded-lg text-sm font-semibold transition-colors ${
-                        settings.defaultActivity === a
-                          ? 'bg-indigo-100 dark:bg-indigo-900/40 text-indigo-600 dark:text-indigo-400 ring-2 ring-indigo-500'
-                          : 'bg-gray-100 dark:bg-gray-800 text-gray-600 dark:text-gray-400 hover:bg-gray-200 dark:hover:bg-gray-700'
-                      }`}
-                    >
-                      {a}
-                    </button>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
+        <SettingsModal
+          isOpen={showSettings}
+          onClose={() => setShowSettings(false)}
+        />
       </main>
     </div>
   )
