+++ src/utils/goalsApi.ts
@@ -1,0 +1,75 @@
+import { Goal } from '@/types/workout'
+
+// API helpers for goals with localStorage fallback
+export const GOALS_STORAGE_KEY = 'fitness-tracker-goals'
+
+export const saveGoalsToStorage = (goals: Goal[]) => {
+  try {
+    if (typeof window === 'undefined') return
+    localStorage.setItem(GOALS_STORAGE_KEY, JSON.stringify(goals))
+  } catch (error) {
+    console.error('Error saving goals to localStorage:', error)
+  }
+}
+
+export const loadGoalsFromStorage = (): Goal[] => {
+  try {
+    if (typeof window === 'undefined') return []
+    const stored = localStorage.getItem(GOALS_STORAGE_KEY)
+    if (stored) {
+      const parsed = JSON.parse(stored)
+      // Convert date strings back to Date objects
+      return parsed.map((goal: Omit<Goal, 'createdAt' | 'updatedAt'> & { createdAt: string; updatedAt: string }) => ({
+        ...goal,
+        createdAt: new Date(goal.createdAt),
+        updatedAt: new Date(goal.updatedAt)
+      }))
+    }
+  } catch (error) {
+    console.error('Error loading goals from localStorage:', error)
+  }
+  return []
+}
+
+// API functions for cross-device persistence
+export const fetchGoalsFromAPI = async (): Promise<Goal[]> => {
+  try {
+    const response = await fetch('/api/goals')
+    if (!response.ok) throw new Error('Failed to fetch goals')
+    const data = await response.json()
+    return data.goals.map((goal: Goal & { createdAt: string; updatedAt: string }) => ({
+      ...goal,
+      createdAt: new Date(goal.createdAt),
+      updatedAt: new Date(goal.updatedAt)
+    }))
+  } catch (error) {
+    console.error('Error fetching goals from API:', error)
+    // Fallback to localStorage
+    return loadGoalsFromStorage()
+  }
+}
+
+export const saveGoalToAPI = async (goalData: Omit<Goal, 'id' | 'createdAt' | 'updatedAt'>, existingId?: string): Promise<Goal | null> => {
+  try {
+    const url = '/api/goals'
+    const method = existingId ? 'PUT' : 'POST'
+    const body = existingId ? { ...goalData, id: existingId } : goalData
+    
+    const response = await fetch(url, {
+      method,
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(body)
+    })
+    
+    if (!response.ok) throw new Error('Failed to save goal')
+    const data = await response.json()
+    return {
+      ...data.goal,
+      createdAt: new Date(data.goal.createdAt),
+      updatedAt: new Date(data.goal.updatedAt)
+    }
+  } catch (error) {
+    console.error('Error saving goal to API:', error)
+    return null
+  }
+}
